zsh: Drop omz and use my own zshrc
I haven't been using anything except the most basic functionality of
OMZ, but I've been suffering the overhead of the entire framework.

This drops it, and moves over to having a custom zshrc with everything.
It's also a lot tidier than it was!

A few plug ins are still there but most of it is hand-customised. This
also allowed for some finer tuning of some options, and allows a lot
more flexibility in future.

It's also just 59 lines of code!