systemd: Don't wait 90s before killing
systemd will wait for it 90 seconds before killing a process that
doesn't exit cleanly by itself.

This makes sense if you want to be conservative or know you have
applications that take a while to shut down (e.g.: an application
server, or a large database, and many other examples I can't think of
right now). On a usual desktop environment, nothing takes more than 2-3
seconds to exit cleanly. Especially if it's just a lone desktop with no
network mounts or alike.

The systemd devs doesn't provide much clarity on how we're supposed to
debug this -- we're just pointed to ask the distribution for help, but
nobody there knows how to debug this either.

This hack just kills any misbehaving process after 3 seconds. Nothing
takes more than 3 seconds to exit on a modern desktop anyway; if
something does, it's probably not gonna exit on its own anyway.