zsh: Remove INC_APPEND_HISTORY
This is mutually exclusive with `SHARE_HISTORY`, as mentioned in the
docs for `INC_APPEND_HISTORY_TIME`[1][2]. This does not fix anything, but
since the options are documented to be mutually exclusive, it's probably
best to remove it.

This also results in no breakage; I probably left it in since it was
needed at the time, and I only added `SHARE_HISTORY` at a later point in
time.

[1]: http://zsh.sourceforge.net/Doc/Release/Options.html
[2]: https://unix.stackexchange.com/questions/592110/zsh-wont-search-through-all-of-my-history