nvim: Install blamer.nvim
Blamer shows a git-blame summary at the end of the current line.
I've reduced to time to zero since otherwise it's odd to see visual
changes happening lagged after switching lines.

This also gives better visual awareness of the cursor current line
(e.g.: it highlights the current line on the right margin too).

The default date format is ambiguous, so I've switched to to ISO-8601,
which is actually readable.