vim: Restore swap file usage
vim-polyglot introduces some changes that break several vim features and
results in data lots in a few rather-common scenarios.

It's unclear what the intent here is, since the related commit does not
describe what the intent is.

See https://github.com/sheerun/vim-polyglot/issues/638

Given that this bug has been around for a month with no attention, it
might be worth dropping and blacklisting vim-polyglot. It's clearly
highly unreliable, and the issues it's introducing are too risky.