waybar: Remove now-breaking workaround
Waybar had an issue in which enable-bar-scroll did not work unless
`smooth-scrolling-threshold=4`. It's unclear **what** this settings did,
but made the feature work.

As of waybar 0.9.5, `enable-bar-scroll` has been fixed and this
workaround is no longer required. Actually, this workaround now breaks
scrolling; the settings `smooth-scrolling-threshold=4` changes waybar's
behaviour so that one has to scroll four times to switch a single
workspace.

It's likely this was the original intent, but the feature has only now
been stabilised.

Related: https://github.com/Alexays/Waybar/issues/367