geoclue-agent: Initial commit
As of recently, geoclue requires an agent to be running to work. This is
likely temporary until Flatpak itself has access control for it. In the
meantime, users outside of Gnome and alike, are required to run an
agent.

The demo agent doesn't do any access control, but suits my purposes just
fine.