waybar: More efficient microphone script
The previous script had a `poll` approach which continuously updated the
state. This was very inefficient, and could also have a short delay.

This version listens to events sent by pulseaudio. This makes the script
sleep 99% of the time, and only wakeup when something needs to be done.

Even so, `grep --line-buffered` intercepts most of the unrelated
messages in a very efficient manner.