fontconfig: Remove superfluous settings
I wrote this file on 2013 (or maybe even earlier) to get the font
antialiasing and hinting I wanted. Over the years, many of fontconfig's
and Arch's defaults have gradually changed, and now all of the defaults
match what I'd picked.

It seems the community has fully agreed on the same preference I had at
the time, so this file is no longer necessary; it just redefined the
default values.