volctl: Rewrite to be more efficient
Use synchronous notifications, so that the notification deamon knows it
has to replace the previous one (and no need to juggle with lock files
and ids).

Also subscribe to pulseaudio events, so that we run only asynchronously
when an event happens. This avoids the need to call it explicitly, but
also picks up other events, like volume changes via pavucontrol, or
hardware buttons.