volctl: Further refine the daemon
Make it a lot less chatty. Logging was pretty verbose, which was fine
for a cli, but is too much for a daemon. It even managed to add some lag
to it.

Also work asynchronously; don't block reading events while handling the
previous one.

Finally, tweak the look and feel to something better integrated.